requirements.txt
torch==2.8.0
torchvision==0.23.0
torchaudio==2.8.0
transformers==4.57.6
gradio==4.44.1
pandas==2.3.3
numpy==2.4.4
scikit-learn==1.8.0
opencv-python-headless==4.13.0.92
Pillow==10.4.0
onnx==1.21.0
onnxruntime==1.24.4
skl2onnx==1.20.0
joblib==1.5.3
matplotlib==3.10.8
shap==0.51.0
xgboost==2.1.4
timm==1.0.26
ultralytics==8.4.33
fastapi==0.112.2
starlette==0.38.6
uvicorn[standard]==0.30.6
python-multipart==0.0.22
